chore(viewer): Move list-view, grid-view and table-view in a subfolder
