+++ projects/knora/viewer/src/lib/view/index.ts
@@ -1,8 +1,8 @@
 export * from './compare-view/compare-view.component';
 export * from './graph-view/graph-view.component';
-export * from './grid-view/grid-view.component';
-export * from './list-view/list-view.component';
+export * from './list/grid-view/grid-view.component';
+export * from './list/list-view/list-view.component';
 export * from './properties-view/properties-view.component';
 export * from './resource-view/resource-view.component';
-export * from './table-view/table-view.component';
+export * from './list/table-view/table-view.component';
 export * from './search-results/search-results.component';
