+++ projects/knora/viewer/src/lib/view/list/grid-view/grid-view.component.ts
