+++ projects/knora/viewer/src/lib/view/list/list-view/list-view.component.html
