+++ projects/knora/viewer/src/lib/view/list/table-view/table-view.component.html
